Backend/requirements.txt
python-multipart==0.0.6
openai==1.46.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
uvicorn
gunicorn
certifi==2023.7.22
pymongo==4.6.0
unstructured
fastapi
tiktoken
pypdf
urllib3
tavily-python
azure-storage-blob
